v0.13.0
- update to chrono 0.4.20.rc.1
- preserve text filter in HLS mode when clips changes
- cross compile ffpapi for macOS

+++ scripts/build_all.sh
@@ -1,13 +1,15 @@
 #!/usr/bin/bash
 
-cd ffplayout-frontend
+if [[ ! -d public ]]; then
+    cd ffplayout-frontend
 
-npm install
-npm run build
-yes | rm -rf ../public
-mv dist ../public
+    npm install
+    npm run build
+    yes | rm -rf ../public
+    mv dist ../public
 
-cd ..
+    cd ..
+fi
 
 targets=("x86_64-unknown-linux-musl" "aarch64-unknown-linux-gnu" "x86_64-pc-windows-gnu" "x86_64-apple-darwin" "aarch64-apple-darwin")
 
@@ -41,11 +43,12 @@ for target in "${targets[@]}"; do
             rm -f "ffplayout-v${version}_${target}.tar.gz"
         fi
 
-        cargo build --release --target=$target --bin ffplayout
+        CC="x86_64-apple-darwin20.4-cc" cargo build --release --target=$target
 
+        cp ./target/${target}/release/ffpapi .
         cp ./target/${target}/release/ffplayout .
-        tar -czvf "ffplayout-v${version}_${target}.tar.gz" --exclude='*.db' assets docs LICENSE README.md ffplayout
-        rm -f ffplayout
+        tar -czvf "ffplayout-v${version}_${target}.tar.gz" --exclude='*.db' assets docs public LICENSE README.md ffplayout ffpapi
+        rm -f ffplayout ffpapi
     else
         if [[ -f "ffplayout-v${version}_${target}.tar.gz" ]]; then
             rm -f "ffplayout-v${version}_${target}.tar.gz"
@@ -55,7 +58,7 @@ for target in "${targets[@]}"; do
 
         cp ./target/${target}/release/ffpapi .
         cp ./target/${target}/release/ffplayout .
-        tar -czvf "ffplayout-v${version}_${target}.tar.gz" --exclude='*.db' assets docs LICENSE README.md ffplayout ffpapi
+        tar -czvf "ffplayout-v${version}_${target}.tar.gz" --exclude='*.db' assets docs public LICENSE README.md ffplayout ffpapi
         rm -f ffplayout ffpapi
     fi
